refactor to react router
