+++ project/src/Styles/Profile.scss
@@ -17,7 +17,7 @@
             font-size: $h3;
         }
     }
-    .followPopUpSection{
+    .followButtons{
         display: flex;
         justify-content: space-around;
         align-items: center;
